improve code

+++ app/src/main/java/com/greybot/mycosts/data/repository/file/FileDataSource.kt
@@ -21,8 +21,8 @@ class FileDataSource @Inject constructor(private val repo: FileRepo) {
 
     suspend fun addFile(
         rowName: String,
-        count: Float = 0f,
-        price: Float = 0F,
+        count: Double = 0.0,
+        price: Double = 0.0,
         currency: CurrencyDto? = null,
         parentId: String?
     ) {
@@ -48,7 +48,7 @@ class FileDataSource @Inject constructor(private val repo: FileRepo) {
         } ?: false
     }
 
-    suspend fun changePrice(objectId: String, count: Float, price: Float) {
+    suspend fun changePrice(objectId: String, count: Double, price: Double) {
         findById(objectId)
             ?.copy(count = count, price = price)
             ?.let { update(it) }
